Reorganize package with abstract classes in objects module
