--- identpy/Error/__init__.py
@@ -1,1 +1,0 @@
-from identpy.Error.WLS_Error import wls_eval
