--- identpy/Method/__init__.py
@@ -1,7 +1,0 @@
-# Abstract class
-from identpy.Method.method import Method
-
-# Methods
-from identpy.Method.MVMO import MVMO
-from identpy.Method.PSO import PSO
-from identpy.Method.TS import TS
