--- identpy/Model/Implicit_Methods/__init__.py
@@ -1,2 +1,0 @@
-from identpy.Model.Implicit_Methods.i_method import IMethod
-from identpy.Model.Implicit_Methods.RK4 import RK4
