--- identpy/Model/__init__.py
@@ -1,8 +1,0 @@
-# Abstract class
-from identpy.Model.model import Model
-
-# Models
-from identpy.Model.systemSM import SpringMass
-from identpy.Model.systemPend import Pendulum
-from identpy.Model.systemZIM import ZIM
-from identpy.Model.systemDFIG import DFIG
