--- identpy/Objects/__init__.py
@@ -1,1 +1,0 @@
-from identpy.Objects.estimator import Estimator
