+++ identpy/error_evaluation/__init__.py
@@ -1,0 +1,1 @@
+from .wls_eval import wls_eval
