+++ identpy/methods/__init__.py
@@ -1,0 +1,4 @@
+# Methods
+from .mvmo import MVMO
+from .pso import PSO
+from .trajectory_sensitivity import TS
