+++ identpy/methods/parameter_classification/__init__.py
@@ -1,0 +1,1 @@
+from .classification_methods import velez_reyes
