+++ identpy/models/implicit_methods/__init__.py
@@ -1,0 +1,1 @@
+from .runge_kutta import RK4
