+++ identpy/objects/__init__.py
@@ -1,0 +1,4 @@
+from .estimator import Estimator
+from .method import Method
+from .implicit_method import ImplicitMethod
+from .model import Model
