+++ tests/test_classification.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 
-from identpy.Method.Parameter_Classification.classification_methods import velez_reyes
+from identpy.methods.parameter_classification import velez_reyes
 
 
 class TestClassification(unittest.TestCase):
